This commit was manufactured by cvs2svn to create tag 'tulip_2_0_6'.
git-svn-id: https://svn.code.sf.net/p/auber/code/tags/tulip_2_0_6@505 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-ogl/include/tulip/ObservableGlGraph.h
@@ -25,7 +25,7 @@ namespace std {
   template<>
   struct less<GlGraphObserver *> {
     size_t operator()(const GlGraphObserver * obs1,const GlGraphObserver *obs2) const {
-      return (unsigned)obs1<(unsigned)obs2;
+      return (unsigned long)obs1<(unsigned long)obs2;
     }
   };
 }
